extract websocket message handler
